+++ tests/pylecroy_tst_01.py
@@ -1,6 +1,6 @@
-from pylecroy.pylecroy import Lecroy
+from pylecroy.pylecroy import *
 import sys
-import logging
+import time
 
 VERSION = '1.0'
 
@@ -11,7 +11,6 @@ Usage:
 Options:
     -h, --help              this help message.
     -v, --version           version info.
-    -l, --logging           enable logging
     -a, --address           device IP address
 '''
 
@@ -22,9 +21,8 @@ def main(argv=None):
     if argv is None:
         argv = sys.argv[1:]
     try:
-        opts, args = getopt.gnu_getopt(argv, 'hvla:', ['help', 'version', 'logging', 'address='])
+        opts, args = getopt.gnu_getopt(argv, 'hva:', ['help', 'version',  'address='])
         address = None
-        log = False
 
         for o, a in opts:
             if o in ('-h', '--help'):
@@ -33,8 +31,6 @@ def main(argv=None):
             if o in ('-v', '--version'):
                 print(VERSION)
                 return 0
-            elif o in ('-l', '--logging'):
-                log = True
             elif o in ('-a', '--address'):
                 address = a
 
@@ -49,30 +45,17 @@ def main(argv=None):
         sys.stderr.write("scope address is mandatory..."+ "\n")
         sys.stderr.write(USAGE + "\n")
 
-    if log is True:
-        logging.basicConfig(level=logging.INFO)
-
     scope = Lecroy(address)
-    # Get identify
-    scope.identify()
-    print("scope identifier   : " + scope.identifier)
+    # Get scope identifier property
+    print("scope identifier : {} ".format(scope.identifier))
 
-    # Test trigger mode
-    scope.set_trigger_mode(scope.AUTO)
-    scope.get_trigger_mode()
-    print("scope trigger mode : " + scope.trigger_mode)
-    scope.set_trigger_mode(scope.NORM)
-    scope.get_trigger_mode()
-    print("scope trigger mode : " + scope.trigger_mode)
-    scope.set_trigger_mode(scope.SINGLE)
-    scope.get_trigger_mode()
-    print("scope trigger mode : " + scope.trigger_mode)
-    scope.set_trigger_mode(scope.STOP)
-    scope.get_trigger_mode()
-    print("scope trigger mode : " + scope.trigger_mode)
+    for trigger in TriggerModes.MODES:
+        scope.trigger_mode = trigger
+        print("scope trigger mode : " + scope.trigger_mode)
+        time.sleep(1)
 
     try:
-        scope.set_trigger_mode("NONE")
+        scope.trigger_mode = "NONE"
     except ValueError as msg:
         print(msg)