+++ tests/pylecroy_tst_02.py
@@ -1,6 +1,6 @@
-from pylecroy.pylecroy import Lecroy
+from pylecroy.pylecroy import *
 import sys
-import logging
+import time
 
 VERSION = '1.0'
 
@@ -11,7 +11,6 @@ Usage:
 Options:
     -h, --help              this help message.
     -v, --version           version info.
-    -l, --logging           enable logging
     -a, --address           device IP address
 '''
 
@@ -22,9 +21,8 @@ def main(argv=None):
     if argv is None:
         argv = sys.argv[1:]
     try:
-        opts, args = getopt.gnu_getopt(argv, 'hvla:', ['help', 'version', 'logging', 'address='])
+        opts, args = getopt.gnu_getopt(argv, 'hva:', ['help', 'version', 'address='])
         address = None
-        log = False
 
         for o, a in opts:
             if o in ('-h', '--help'):
@@ -33,8 +31,6 @@ def main(argv=None):
             if o in ('-v', '--version'):
                 print(VERSION)
                 return 0
-            elif o in ('-l', '--logging'):
-                log = True
             elif o in ('-a', '--address'):
                 address = a
 
@@ -49,46 +45,43 @@ def main(argv=None):
         sys.stderr.write("scope address is mandatory..."+ "\n")
         sys.stderr.write(USAGE + "\n")
 
-    if log is True:
-        logging.basicConfig(level=logging.INFO)
-
     scope = Lecroy(address)
 
-    # Test identify
-    scope.identify()
-    print("scope identifier   : " + scope.identifier)
+    # Get scope identifier property
+    print("scope identifier : {} ".format(scope.identifier))
+    # Stop Trigger
+    scope.trigger_mode = TriggerModes.STOP
 
-    # print all C1 parameters
-    print("Channel C1 Parameters : ", scope.get_parameter(scope.C1, "ALL"))
+    input("Get a signal on C1 and press a key to continue...")
 
     # grid state
-    for state in scope.GRID_STATE:
+    for state in GridStates.STATES:
         print("Grid state : " + state)
-        scope.show_grid(state)
-        input("Press a key to continue...")
-    # single grid
-    scope.show_grid(scope.SINGLE)
+        scope.grid = state
+        time.sleep(1)
+
+    # Dual grid
+    scope.grid = GridStates.DUAL
 
     # Set  C1 to C4 to off
     print("Channels  OFF...")
-    for channel in scope.WAVEFORM_CHANNELS:
-        scope.show_trace(channel, "OFF")
+    for channel in Channels.NAMES:
+        scope.display_channel(channel, "OFF")
     print("Memory  OFF...")
-    for channel in scope.INTERNAL_MEMORY:
-        scope.show_trace(channel, "OFF")
+    for Memory in Memories.NAMES:
+        scope.display_channel(Memory, "OFF")
 
-    input("Press a key to continue...")
     print("Channel C1 ON...")
-    scope.show_trace(scope.C1, "ON")
+    scope.display_channel(Channels.C1, "ON")
 
-    input("Get a signal on C1 and press a key to continue...")
+    print("Set Trigger Single...")
+    scope.trigger_mode = TriggerModes.SINGLE
+
     # save C1 to M1
     print("Channel C1 save in M1...")
-    scope.save_memory(scope.C1, scope.M1)
-    print("Channel C1 OFF...")
-    scope.show_trace(scope.C1, "OFF")
+    scope.save_memory(Channels.C1, Memories.M1)
     print("Show M1...")
-    scope.show_trace(scope.M1, "ON")
+    scope.display_channel(Memories.M1, "ON")
 
     scope.close()